restructure workflow manager folder
