+++ react-components/oodt_opsui_sample_app/src/components/workflowManager/index.js
@@ -17,4 +17,6 @@
 
 // Export all the explicitly exported components, this file will contain our
 // components when built by webpack and sent off to the world.
-export * from "./components";
+import WorkflowList from "./WorkflowList";
+
+export { WorkflowList };
